pyproject.toml
[project]
name = "bookhoob"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikhail",email = "miplik@yandex.ru"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.7"
djangorestframework = "^3.16.1"
boto3 = "^1.40.69"
django-storages = "^1.14.6"
django-environ = "^0.12.0"
pillow = "^12.0.0"
drf-spectacular = "^0.29.0"
djangorestframework-simplejwt = "^5.5.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit=">=4.4.0,<5.0.0"
pytest = "^9.0.1"
ruff = "^0.14.5"

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
]
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "BLE",
    "C90",
    "E",
    "F",
    "G",
    "I",
    "ICN",
    "ISC",
    "PLE",
    "Q",
    "RUF006",
    "RUF100",
    "T10",
    "T20",
    "TID",
    "UP",
    "W",
]
ignore = ["ISC001"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["book_hub", "tests"]
